+++ internal/pkg/instrumentation/bpf/github.com/segmentio/kafka-go/consumer/probe_test.go
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gin
+package consumer
 
 import (
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
+
 	"go.opentelemetry.io/otel/attribute"
 	semconv "go.opentelemetry.io/otel/semconv/v1.24.0"
 	"go.opentelemetry.io/otel/trace"
@@ -40,10 +41,14 @@ func TestProbeConvertEvent(t *testing.T) {
 			EndTime:     uint64(end.UnixNano()),
 			SpanContext: context.EBPFSpanContext{TraceID: traceID, SpanID: spanID},
 		},
-		// "GET"
-		Method: [7]byte{0x47, 0x45, 0x54},
-		// "/foo/bar"
-		Path: [100]byte{0x2f, 0x66, 0x6f, 0x6f, 0x2f, 0x62, 0x61, 0x72},
+		// topic1
+		Topic: [256]byte{0x74, 0x6f, 0x70, 0x69, 0x63, 0x31},
+		// key1
+		Key: [256]byte{0x6b, 0x65, 0x79, 0x31},
+		// test consumer group
+		ConsumerGroup: [128]byte{0x74, 0x65, 0x73, 0x74, 0x20, 0x63, 0x6f, 0x6e, 0x73, 0x75, 0x6d, 0x65, 0x72, 0x20, 0x67, 0x72, 0x6f, 0x75, 0x70},
+		Offset:        42,
+		Partition:     12,
 	})
 
 	sc := trace.NewSpanContext(trace.SpanContextConfig{
@@ -52,14 +57,19 @@ func TestProbeConvertEvent(t *testing.T) {
 		TraceFlags: trace.FlagsSampled,
 	})
 	want := &probe.SpanEvent{
-		SpanName:    "GET",
+		SpanName:    kafkaConsumerSpanName("topic1"),
 		StartTime:   int64(start.UnixNano()),
 		EndTime:     int64(end.UnixNano()),
 		SpanContext: &sc,
 		Attributes: []attribute.KeyValue{
-			semconv.HTTPRequestMethodKey.String("GET"),
-			semconv.URLPath("/foo/bar"),
+			semconv.MessagingSystemKafka,
+			semconv.MessagingOperationReceive,
+			semconv.MessagingKafkaDestinationPartition(12),
+			semconv.MessagingDestinationName("topic1"),
+			semconv.MessagingKafkaMessageOffset(42),
+			semconv.MessagingKafkaMessageKey("key1"),
+			semconv.MessagingKafkaConsumerGroup("test consumer group"),
 		},
 	}
-	assert.Equal(t, want, got)
+	assert.Equal(t, want, got[0])
 }